feat: Add AcceptLoan and RejectLoan methods to loan use case
This commit adds the AcceptLoan and RejectLoan methods to the loan use case in loan_usecase.go. These methods allow the loan use case to accept or reject a loan by interacting with the loan repository. This change enhances the functionality of the loan use case and provides more control over loan processing.